# Tokenizer: Optimized tokenization with exceptions