--- a/src/test/scala/com/johnsnowlabs/nlp/annotators/TokenizerTestSpec.scala
+++ b/src/test/scala/com/johnsnowlabs/nlp/annotators/TokenizerTestSpec.scala
@@ -18,12 +18,15 @@ package com.johnsnowlabs.nlp.annotators
 
 import com.johnsnowlabs.nlp._
 import com.johnsnowlabs.nlp.annotators.sbd.pragmatic.SentenceDetector
-import com.johnsnowlabs.tags.FastTest
+import com.johnsnowlabs.tags.{FastTest, SlowTest}
+import com.johnsnowlabs.util.Benchmark
 import org.apache.spark.ml.{Pipeline, PipelineModel, Transformer}
 import org.apache.spark.sql.{DataFrame, Dataset, Row}
 import org.scalatest.flatspec.AnyFlatSpec
 
 import java.util.Date
+import scala.collection.mutable
+import scala.collection.mutable.ArrayBuffer
 
 class TokenizerTestSpec extends AnyFlatSpec with TokenizerBehaviors {
 
@@ -253,9 +256,9 @@ class TokenizerTestSpec extends AnyFlatSpec with TokenizerBehaviors {
     Annotation(AnnotatorType.TOKEN, 60, 60, ".", Map("sentence" -> "0")))
 
   def getTokenizerOutput[T](
-      tokenizer: TokenizerModel,
-      data: DataFrame,
-      mode: String = "finisher"): Array[T] = {
+                             tokenizer: TokenizerModel,
+                             data: DataFrame,
+                             mode: String = "finisher"): Array[T] = {
     val finisher = new Finisher()
       .setInputCols("token")
       .setOutputAsArray(true)
@@ -272,10 +275,10 @@ class TokenizerTestSpec extends AnyFlatSpec with TokenizerBehaviors {
 
   /* here assembler can either be a SentenceDetector or a DocumentAssembler */
   def getTokenizerPipelineOutput[T](
-      assembler: Transformer,
-      tokenizer: TokenizerModel,
-      data: DataFrame,
-      mode: String = "finisher"): Array[T] = {
+                                     assembler: Transformer,
+                                     tokenizer: TokenizerModel,
+                                     data: DataFrame,
+                                     mode: String = "finisher"): Array[T] = {
 
     val finisher = new Finisher()
       .setInputCols("token")
@@ -658,6 +661,51 @@ class TokenizerTestSpec extends AnyFlatSpec with TokenizerBehaviors {
       s"because result tokens differ: " +
         s"\nresult was \n${result.mkString("\n")} \nexpected is: \n${expected.mkString("\n")}")
   }
+
+  "a Tokenizer" should "benchmark exceptions" taggedAs SlowTest in {
+    val data = AnnotatorBuilder.getTrainingDataSet("src/test/resources/spell/sherlockholmes.txt")
+
+    val documentAssembler = new DocumentAssembler()
+      .setInputCol("text")
+      .setOutputCol("document")
+
+    val tokenizer = new Tokenizer()
+      .setInputCols("document")
+      .setOutputCol("token")
+
+    val pipeline = new Pipeline().setStages(Array(
+      documentAssembler,
+      tokenizer
+    ))
+
+    /** Run first to cache for more consistent results */
+    var result: Array[Row] = pipeline.fit(data).transform(data).select("token.result").collect()
+
+    val tokens = result.foldLeft(ArrayBuffer.empty[String]) { (arr: ArrayBuffer[String], i: Row) =>
+      val Row(tokens: mutable.WrappedArray[String]) = i
+      arr ++= tokens.map(_.replaceAll("\\W", "")).filter(_.nonEmpty)
+    }.toArray
+
+    val documentAssembler2 = new DocumentAssembler()
+      .setInputCol("text")
+      .setOutputCol("document")
+
+    val tokenizerWithExceptions = new Tokenizer()
+      .setInputCols("document")
+      .setOutputCol("token")
+      .setExceptions(tokens.slice(0, 200))
+
+    val pipelineWithExceptions = new Pipeline().setStages(Array(
+      documentAssembler2,
+      tokenizerWithExceptions
+    ))
+
+    Benchmark.measure(iterations=20, forcePrint = true, description = "Time to tokenize Sherlock Holmes with exceptions") {
+      pipelineWithExceptions.fit(data).transform(data).select("token.result").collect()
+    }
+
+  }
+
 
   "RecursiveTokenizer" should "split suffixes" taggedAs FastTest in {