<!--- Provide a general summary of your changes in the Title above -->

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
The current Tokenizer implementation has performance issues if a large number of exceptions are used. 

For example:
When tokenizing `src/test/resources/spell/sherlockholmes.txt`, on my machine it will take an average of 0.58 seconds for 20 iterations. If there are 100 exceptions, it will take on average 3.17 seconds, for 200 it will jump to 5.8 seconds. The following benchmark was used for testing: [TokenizerTestSpec](https://github.com/DevinTDHa/spark-nlp/blob/c2506ceb90ee33998e02449b9f6f623523972092/src/test/scala/com/johnsnowlabs/nlp/annotators/TokenizerTestSpec.scala#L665)

There is double the compilations of regular expressions and a severe performance drop. 

### Problem

The issue lies with how the exceptions are checked. 
1. Regex are compiled for each sentence that is tagged. This can be done beforehand.
    - Addtionally, each regex exception is compiled twice due to first being compiled to replace the `BREAK_PATTERN` in [`tag`](https://github.com/JohnSnowLabs/spark-nlp/blob/24524a7fbb73fb2411879e062fd33b89b3dddbb1/src/main/scala/com/johnsnowlabs/nlp/annotators/TokenizerModel.scala#L244) and afterwards compiled again in [`casedMatchExists`](https://github.com/JohnSnowLabs/spark-nlp/blob/24524a7fbb73fb2411879e062fd33b89b3dddbb1/src/main/scala/com/johnsnowlabs/nlp/annotators/TokenizerModel.scala#L226-L228)

## Description
<!--- Describe your changes in detail -->


The following changes were implemented:
  - exceptions and regular expressions are compiled during instantiation of the model instead of for each sentence
  - exceptions are combined into one regular expression in [`TokenizerModel`](https://github.com/DevinTDHa/spark-nlp/blob/c2506ceb90ee33998e02449b9f6f623523972092/src/main/scala/com/johnsnowlabs/nlp/annotators/TokenizerModel.scala#L262-L277)
  - function `casedMatchExists` was removed in favor of a hash set that checks whether an exception without a break exists in [`tag`](https://github.com/DevinTDHa/spark-nlp/blob/c2506ceb90ee33998e02449b9f6f623523972092/src/main/scala/com/johnsnowlabs/nlp/annotators/TokenizerModel.scala#L292-L313)
    - @maziyarpanahi Do you think this approach is acceptable?
  - fixed some minor warnings highlighted by intelliJ

### Benchmark Results
These changes will result in a performance boost of 
  - more than 300% for 100 exceptions (from 3.17 down to 0.95 seconds) and 
  - for 200 exceptions a performance boost of more than 450% (from 5.8 seconds down to 1.24 seconds).

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Added an additional test for the benchmark. All Tokenizer tests are passing.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Code improvements with no or little impact
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the [CONTRIBUTING](http://nlp.johnsnowlabs.com/contribute.html) page.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.