@DevinTDHa 2022-04-11
@maziyarpanahi I will test how well it performs. In this particular case, the hashset will not grow much: at most it will have nr. of exceptions entries and only if these exceptions are not affected by the `BREAK_PATTERN`.