@DevinTDHa 2022-04-14
@maziyarpanahi I looked at it the following ways:

I used free samples form the COCA Dataset as a test. In total these are about 500k sentences, 56 million characters.

As exceptions I chose something ridiculous like `\w+`, which will effectively match every word in a sentence. I also monitored the tests with visualvm.

- Tokenizing it the reuglar way has no issues. The set doesn't grow beyond the vocabulary of the sentence and the set also only exists for each sentence, keeping it small.
- For the second test, I replaced all `\n` with `\t` in the sample data. This effectively results in one massive sentence, which will be tokenized at once. The collection will grow to a fixed size of the vocabulary of the data (which only happens once). The processing of the exceptions will be done in about 5 seconds on my machine. 
  - **Note:** With the current Tokenizer Implementation, the whole tokenization process doesn't terminate (at least I didn't wait for that long). Followingly, the current implementation of the exceptions will perform even worse.